+++ src/context/FilesDialogsContext.tsx
@@ -4,6 +4,8 @@ import { File } from "../types/entities";
 export enum AVAILABLE_DIALOGS {
   "CREATE_FOLDER" = "CREATE_FOLDER",
   "RENAME_FILE" = "RENAME_FILE",
+  "UPLOAD_FILE" = "UPLOAD_FILE",
+  "DELETE_FILE" = "DELETE_FILE",
   "ACCESS_MANAGEMENT" = "ACCESS_MANAGEMENT",
   "MOVE_FILE" = "MOVE_FILE"
 }
@@ -19,8 +21,10 @@ const defaultValues: FilesDialogsContext = {
   dialogsVisibilityState: {
     [AVAILABLE_DIALOGS.CREATE_FOLDER]: false,
     [AVAILABLE_DIALOGS.RENAME_FILE]: false,
+    [AVAILABLE_DIALOGS.UPLOAD_FILE]: false,
     [AVAILABLE_DIALOGS.ACCESS_MANAGEMENT]: false,
-    [AVAILABLE_DIALOGS.MOVE_FILE]: false
+    [AVAILABLE_DIALOGS.MOVE_FILE]: false,
+    [AVAILABLE_DIALOGS.DELETE_FILE]: false
   },
   selectedFile: null,
   openDialog: () => {},
@@ -40,6 +44,8 @@ export const FilesDialogsContextProvider = ({
   >({
     [AVAILABLE_DIALOGS.CREATE_FOLDER]: false,
     [AVAILABLE_DIALOGS.RENAME_FILE]: false,
+    [AVAILABLE_DIALOGS.UPLOAD_FILE]: false,
+    [AVAILABLE_DIALOGS.DELETE_FILE]: false,
     [AVAILABLE_DIALOGS.ACCESS_MANAGEMENT]: false,
     [AVAILABLE_DIALOGS.MOVE_FILE]: false
   });