+++ src/screens/FilePage.tsx
@@ -9,7 +9,9 @@ import { useContext } from "react";
 import {
   CreateFolderDialog,
   EditNameDialog,
+  UploadFileDialog,
   AccessManagementDialog,
+  DeleteFileDialog,
   MoveFileDialog
 } from "./dialogs/index";
 
@@ -21,6 +23,9 @@ export function FilePage() {
 
   const showAccessDialog =
     dialogsVisibilityState[AVAILABLE_DIALOGS.ACCESS_MANAGEMENT];
+
+  const showDeleteDialog =
+    dialogsVisibilityState[AVAILABLE_DIALOGS.DELETE_FILE];
 
   const showMoveDialog = dialogsVisibilityState[AVAILABLE_DIALOGS.MOVE_FILE];
 
@@ -51,9 +56,11 @@ export function FilePage() {
           )}
         </div>
       </main>
+      <UploadFileDialog />
       <CreateFolderDialog />
       {showRenameDialog && <EditNameDialog />}
       {showAccessDialog && <AccessManagementDialog />}
+      {showDeleteDialog && <DeleteFileDialog />}
       {showMoveDialog && <MoveFileDialog />}
     </div>
   );