+++ src/screens/dialogs/access-file-dialog.tsx
@@ -1,12 +1,14 @@
 import { useContext, useState, useEffect } from "react";
 import { SharedWithWhoService } from "../../services/files/get-users-file.service";
 import { Dialog } from "../../components/Dialog";
+import { shareFileService } from "../../services/files/share-file.service";
 import {
   FilesDialogsContext,
   AVAILABLE_DIALOGS,
   AuthContext
 } from "../../context/index";
 import toast from "react-hot-toast";
+import { unshareFileService } from "../../services/files/unshare-file.service";
 
 export const AccessManagementDialog = () => {
   const [usersWithAccess, setUsersWithAccess] = useState<string[]>([]);
@@ -17,6 +19,23 @@ export const AccessManagementDialog = () => {
   const Open = dialogsVisibilityState[AVAILABLE_DIALOGS.ACCESS_MANAGEMENT];
 
   const { session } = useContext(AuthContext);
+
+  const handleShare = async () => {
+    const shareRequest = {
+      token: session?.token as string,
+      fileUUID: selectedFile?.uuid as string,
+      otherUsername: newAccess
+    };
+    const { success, msg } = await shareFileService(shareRequest);
+    if (!success) {
+      toast.error(msg);
+      return;
+    }
+
+    setNewAccess("");
+    toast.success(msg);
+    setUsersWithAccess([...usersWithAccess, newAccess]);
+  };
 
   useEffect(() => {
     const fetchUsersWithAccess = async () => {
@@ -36,9 +55,24 @@ export const AccessManagementDialog = () => {
     }
   }, [selectedFile]);
 
-  const handleSave = async () => {};
+  const handleUnshare = async (userName: string) => {
+    const unShareRequest = {
+      token: session?.token as string,
+      fileUUID: selectedFile?.uuid as string,
+      otherUsername: userName
+    };
 
-  const handleUnshare = async () => {};
+    const { success, msg } = await unshareFileService(unShareRequest);
+    if (!success) {
+      toast.error(msg);
+      return;
+    }
+
+    setUsersWithAccess(
+      usersWithAccess.filter((username) => username !== userName)
+    );
+    toast.success(msg);
+  };
 
   return (
     <Dialog
@@ -55,8 +89,9 @@ export const AccessManagementDialog = () => {
         className="mb-3 w-full rounded-lg border p-2"
       />
       <button
+        id="share-file"
         className="hover-bg-blue-700 mt-3 rounded-md bg-blue-600 px-4 py-2 text-white"
-        onClick={handleSave}
+        onClick={handleShare}
       >
         Share
       </button>
@@ -69,7 +104,7 @@ export const AccessManagementDialog = () => {
                 <span>{user}</span>
                 <button
                   className="hover-bg-red-600 mt-0.5 rounded-md bg-red-500 px-2 py-1 text-white"
-                  onClick={() => handleUnshare()}
+                  onClick={() => handleUnshare(user)}
                 >
                   Un-share
                 </button>