+++ src/screens/dialogs/index.ts
@@ -1,4 +1,6 @@
 export { CreateFolderDialog } from "./create-folder-dialog";
+export { UploadFileDialog } from "./upload-file-dialog";
 export { EditNameDialog } from "./rename-file-dialog";
 export { AccessManagementDialog } from "./access-file-dialog";
+export { DeleteFileDialog } from "./delete-file-dialog";
 export { MoveFileDialog } from "./move-file-dialog";